Isolate SavePanicTrace() for windows build
